backend/persona-generator/pyproject.toml
[tool.poetry]
name = "persona-generator"
version = "0.1.0"
description = ""
authors = ["devin-ai-integration[bot] <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.6"}
psycopg = {extras = ["binary"], version = "^3.2.3"}
openai = "^1.57.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
